# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API к ИС Не-Ной",
        "title": "Ne-Noy-Api",
        "contact": {},
        "version": "1.0"
    },
    "host": "https://simply-funny-spearfish.cloudpub.ru",
    "basePath": "/api",
    "paths": {
        "/v1/attachments": {
            "get": {
                "security": [
                    {
                        "VkAuth": []
                    }
                ],
                "description": "Возвращает список подписанных временных ссылок для переданных UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attachments"
                ],
                "summary": "Получить временные ссылки на несколько файлов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор запроса",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Список UUID и TTL",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GetManyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetManyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "VkAuth": []
                    }
                ],
                "description": "Передаёт файл в сервис вложений через gRPC и возвращает UUID созданной записи",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attachments"
                ],
                "summary": "Загрузить один файл",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор запроса",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные файла",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PutOneRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.PutOneResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные входные данные",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "VkAuth": []
                    }
                ],
                "description": "Удаляет список файлов по UUID. При force=false физически удаляет планировщик",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attachments"
                ],
                "summary": "Удалить несколько файлов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор запроса",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Список UUID и флаг force",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteManyRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/attachments/batch": {
            "post": {
                "security": [
                    {
                        "VkAuth": []
                    }
                ],
                "description": "Передаёт список файлов в сервис вложений через gRPC и возвращает список UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attachments"
                ],
                "summary": "Загрузить несколько файлов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор запроса",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Список файлов",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PutManyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.PutManyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/attachments/{id}": {
            "get": {
                "security": [
                    {
                        "VkAuth": []
                    }
                ],
                "description": "Возвращает подписанную временную ссылку для скачивания файла по его UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attachments"
                ],
                "summary": "Получить временную ссылку на файл",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор запроса",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID файла",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Время жизни ссылки в секундах (по умолчанию 3600)",
                        "name": "ttl",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetOneResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "VkAuth": []
                    }
                ],
                "description": "Удаляет файл по UUID. При force=false счётчик уменьшается; физически удаляет планировщик",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attachments"
                ],
                "summary": "Удалить файл",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор запроса",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID файла",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Принудительное немедленное удаление (по умолчанию false)",
                        "name": "force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/file": {
            "get": {
                "description": "Возвращает бинарное содержимое файла. Не требует авторизации — доступ контролируется подписью",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "attachments"
                ],
                "summary": "Скачать файл по подписанной ссылке",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID временной ссылки",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Строка подписи",
                        "name": "sign",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Подпись недействительна",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.DeleteManyRequest": {
            "type": "object",
            "properties": {
                "force": {
                    "type": "boolean",
                    "example": false
                },
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetManyRequest": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ttl": {
                    "type": "integer",
                    "example": 3600
                }
            }
        },
        "dto.GetManyResponse": {
            "type": "object",
            "properties": {
                "urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dto.GetOneResponse": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string",
                    "example": "https://storage.example.com/file?sign=abc"
                }
            }
        },
...
        "dto.PutManyRequest": {
            "type": "object",
            "properties": {
                "async": {
                    "type": "boolean",
                    "example": false
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UploadFileDTO"
                    }
                }
            }
        },
        "dto.PutManyResponse": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dto.PutOneRequest": {
            "type": "object",
            "properties": {
                "async": {
                    "type": "boolean",
                    "example": false
                },
                "file": {
                    "$ref": "#/definitions/dto.UploadFileDTO"
                }
            }
        },
        "dto.PutOneResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "dto.StorageType": {
            "type": "integer",
            "format": "int32",
            "enum": [
                1,
                2
            ],
            "x-enum-varnames": [
                "StorageTypeDatabase",
                "StorageTypeObjectStorage"
            ]
        },
        "dto.UploadFileDTO": {
            "type": "object",
            "properties": {
                "file_content": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "file_name": {
                    "type": "string",
                    "example": "photo.png"
                },
                "storage_type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.StorageType"
                        }
                    ],
                    "example": 2
                }
            }
        },
...
    },
    "securityDefinitions": {
        "VkAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}